renamed files to get rid of hyphens, added a few more websites, and polished the github presentation
