+++ src/components/email/EmailTemplate.tsx
